--- backend/.env.example
@@ -1,8 +1,0 @@
-# Connect to Supabase via connection pooling
-DATABASE_URL="postgresql://postgres.PROJECT:[YOUR-PASSWORD]@aws-0-[REGION].pooler.supabase.com:6543/postgres?pgbouncer=true"
-
-# Direct connection to the database. Used for migrations
-DIRECT_URL="postgresql://postgres.PROJECT:[YOUR-PASSWORD]@aws-0-[REGION].pooler.supabase.com:5432/postgres"
-
-# Server Configuration
-PORT=1234
