+++ backend/src/routes/index.ts
@@ -1,0 +1,10 @@
+import { Router } from 'express';
+import { ping, dbTest, serveSwagger } from '../controllers/health.ts';
+
+const router = Router();
+
+router.get('/api/ping', ping);
+router.get('/api/db-test', dbTest);
+router.get('/swagger-output.json', serveSwagger);
+
+export default router;
