+++ backend/src/services/db.ts
@@ -1,0 +1,13 @@
+import { PrismaClient } from '@prisma/client';
+
+export const prisma = new PrismaClient();
+
+export const connectDatabase = async () => {
+  try {
+    await prisma.$connect();
+    console.log('Connected to Supabase PostgreSQL via Prisma');
+  } catch (error) {
+    console.error('Database connection error:', error);
+    process.exit(1);
+  }
+};
